requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
clarifai